[FIX]: Build on windows

+++ client/src/systems/Controls_systems.cpp
@@ -189,7 +189,7 @@ namespace ecs
             std::string back_path = ASSET_FILE("backgrounds/menu/setting_back.jpg");
             ImageComponent setting_back(back_path, SETTINGS_FOCUS,
                                         [](int screenWidth, int screenHeight) {return 0;},
-                                        [](int screenWidth, int screenHeight) {return 0;}, GetScreenWidth(), GetScreenHeight());
+                                        [](int screenWidth, int screenHeight) {return 0;}, static_cast<float>(GetScreenWidth()), static_cast<float>(GetScreenHeight()));
             ecs.add_component<SettingsComponent>(settingEntity, {setting_back});
             ecs.run_event(ChangeFocusEvent{SETTINGS_FOCUS});
         }
